feat: DeepSeek-V4-Flash support on Kunlun XPU
- Post-import hook system replacing fragile module mappings
- Fix inverse RoPE: NeoX half-split -> GPT-J interleaved (is_neox_style=False)
- Handle scalar weight_scale_inv in o_proj dequant
- Add _bf16_weight_cache for o_proj (avoid per-step CPU fp8->bf16 dequant)
- dequantize_fp8_blocks: on-device scale multiply to reduce PCIe traffic
- Register compressor hooks (save_partial_states, compress_norm_rope_store_triton)
- KunlunFp8MoEMethod correctness fallback with CPU routing metadata
- FP8/FP8-block kernel classes for OOT platform
- bf16 cache insert fallback in attention hook

+++ vllm_kunlun/ops/attention/flashmla.py
@@ -234,8 +234,8 @@ def flash_mla_sparse_prefill(
     s_q, h_q, d_qk = q.shape
 
     out = torch.zeros([s_q, h_q, d_v], dtype=q.dtype, device=q.device)
-    max_logits = torch.zeros([s_q, h_q], dtype=torch.float32, device=q.device)
-    lse = torch.zeros([s_q, h_q], dtype=torch.float32, device=q.device)
+    max_logits = torch.zeros([s_q, h_q], dtype=q.dtype, device=q.device)
+    lse = torch.zeros([s_q, h_q], dtype=q.dtype, device=q.device)
 
     torch.ops._C.sparse_prefill_fwd_opt(
         q=q,
@@ -257,7 +257,7 @@ def flash_mla_sparse_prefill(
     # out_scale = 1 / math.log2(math.e)
     # gpu_max_logits * out_scale = kunlun_lse
     # gpu_lse * out_scale = kunlun_lse
-    return out, max_logits, lse
+    return out, max_logits.float(), lse.float()
 
 
 #
